Move RegionAdministrationService to its folder
