+++ EPlast/EPlast.BLL/Services/FormerMember/FormerMemberAdminService.cs
@@ -5,6 +5,7 @@ using EPlast.BLL.Interfaces.FormerMember;
 using EPlast.BLL.Interfaces.GoverningBodies;
 using EPlast.BLL.Interfaces.GoverningBodies.Sector;
 using EPlast.BLL.Interfaces.Region;
+using EPlast.BLL.Interfaces.RegionAdministrations;
 using EPlast.DataAccess.Entities;
 using Microsoft.AspNetCore.Identity;
