+++ EPlast/EPlast.Tests/Services/FormerMember/FormerMemberAdminServiceTest.cs
@@ -5,6 +5,7 @@ using EPlast.BLL.Interfaces.Club;
 using EPlast.BLL.Interfaces.GoverningBodies;
 using EPlast.BLL.Interfaces.GoverningBodies.Sector;
 using EPlast.BLL.Interfaces.Region;
+using EPlast.BLL.Interfaces.RegionAdministrations;
 using EPlast.BLL.Services.FormerMember;
 using EPlast.DataAccess.Entities;
 using Microsoft.AspNetCore.Identity;
