+++ EPlast/EPlast.Tests/Services/Regions/RegionAdministrationServiceTests.cs
@@ -8,6 +8,7 @@ using EPlast.BLL.DTO.Admin;
 using EPlast.BLL.DTO.Region;
 using EPlast.BLL.Interfaces.Admin;
 using EPlast.BLL.Services.Region;
+using EPlast.BLL.Services.RegionAdministrations;
 using EPlast.DataAccess.Entities;
 using EPlast.DataAccess.Repositories;
 using EPlast.Resources;
